requirements.txt
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
celery==5.2.7
django-redis==5.2.0